requirements.txt
streamlit
pandas
python-dateutil
google-generativeai